Trying to get Vasari Loyalist's BuffPlanetLoss to work on planet scuttle
The Stripped to the Core research is supposed to apply BuffPlanetLoss to all ships after *any* sort of planet death, but it seems to only apply to cultural & bombardment-based decolonizations. Scuttling the planet does not apply it.
Current workaround: BuffStrippedToTheCoreSelf will now apply BuffPlanetLoss to any units in orbit around the scuttled planet and any units in orbit around planets adjacent to the scuttled planet. Attempts to chain this effect so units around all planets (thus mimicking the desired effect of planet loss affecting all combat units) causes the game to crash. So this may be the most workable solution.