Fixing range errors
Fixing the (most likely) source of range errors in the debug file, which might ultimately be responsible for crashing the game sometimes. (Vasari Rebel does produce the most of these errors by far.) The cause appears to be that some abilities/buffs featured multiple ability levels, but the range was specified for only one level.