FIXED Vasari Rebel crashing the game
The problem was that the Vasari Rebel titan's ability was trying to check for "number of enemies in range" while I never specified a range. This would spam the debug log with dozens of range errors per second and eventually crash the game. This has been fixed.

Also corrected a completely different buff but that's not as important.